% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_methods.R
\name{components.polcax}
\alias{components.polcax}
\title{Access Components of the polcax Object}
\usage{
\method{components}{polcax}(
  object,
  type = c("data", "class_names", "variables", "class_numbers", "prior_p", "posterior",
    "assignment", "fit_stats", "predcell", "probs", "probs.se"),
  ...
)
}
\arguments{
\item{object}{a \code{polcax} class object.}

\item{type}{component to be retrieved, see \code{Details}.}

\item{...}{extra arguments for downstream functions
like \code{\link{nodal_vote}}, see \code{Details}.}
}
\value{
a tibble or a list with components as specified in \code{Details}.
}
\description{
Accesses/extracts components of the \code{polcax} object.
}
\details{
By specifying the \code{type} arguments, the following components are retrieved:
\itemize{
\item \code{data} the modeling data
\item \code{class_names} names of the classes
\item \code{variables} names of the modeling variables
\item \code{class_numbers} numbers, fraction and percentages of observations assigned
to the classes in a tibble. The class assignment is provided
by \code{\link{nodal_vote}}
\item \code{prior_p} a tibble with mixing proportions (mean of priors) in the
LCA model with their standard deviations.
\item \code{posterior} a class \code{lca_assign} tibble with posterior class membership
probabilities and class assignment (nodal) computed
by \code{\link{nodal_vote}}. See also: \code{\link{plot.lca_assign}}
\item \code{assignment} as above
\item \code{fit_stats} a tibble with fit statistics as specified by
\code{\link{summary.polcax}}
\item \code{predcell} a tibble of observed versus predicted counts of observations.
\item \code{probs} a list of tibbles with estimated class-conditional
response probabilities.
\item \code{probs.se} a list of tibble with standard errors
of the conditional probabilities.
}
}
