% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3step.R
\name{d3step}
\alias{d3step}
\alias{r3step}
\title{Bootstraped 3-step testing of latent class analysis}
\usage{
d3step(formula, x, nclass = NULL, nreps = 1000, alpha = 0.05, ...)

r3step(formula, x, nclass = NULL, nreps = 1000, alpha = 0.05, ...)
}
\arguments{
\item{formula}{Either a vector containing a string indicating the dependent variable to test or an object of class "formula". The details of model specification are given under ‘Details’}

\item{x}{Object of class inheriting from "poLCA" or "poLCA2".}

\item{nclass}{f \code{object} contains many "poLCA" object (i.e., if it a "poLCA2" object), choose the model with the given number of classes.}

\item{nreps}{The number of replications for the bootstrap. Default is 1000.}

\item{alpha}{For the confidence interval. Default if 5\%, of\code{alpha = .05}}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A list containing the comparison indices in a vector and an array containing the conditional median and confidence interval of each classes.
}
\description{
Bootstraped 3-step testing of latent class analysis
}
\details{
Some examples of formula :
\itemize{
  \item  \code{DV ~ 1} is the same as \code{"DV"} or \code{DV ~ classes}.
  \item Add covariate as \code{DV ~ IV} or \code{DV ~ IV + classes}
}
d3step and r3step is the same function. It will detect automatically if \code{DV} is a factor (d3step) or continuous (r3step).
}
\examples{
f1 <- cbind(V1, V2, V3, V4, V5, V6) ~ 1
out <- poLCA(f1, nclass = 1:3, data = ex1.poLCA)
r3step("continuous", out, nclass = 3)
d3step("categorical", out, nclass = 3)
}
