% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex1.poLCA.R
\docType{data}
\name{ex1.poLCA}
\alias{ex1.poLCA}
\title{An example for poLCAExtra}
\format{
A 800 by 8 data frame. The true number of classes is 3 based on 6 indicator. The model is
f1 <- as.formula(cbind(V1, V2, V3, V4, V5, V6) ~ 1). There are two additionnal variable for 3-step testing.
}
\source{
\url{https://github.com/quantmeth}
}
\usage{
ex1.poLCA
}
\description{
A data frame for poLCA and poLCAExtra.
}
\keyword{datasets}
