% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex2.poLCA.R
\docType{data}
\name{ex2.poLCA}
\alias{ex2.poLCA}
\title{An example for poLCAExtra}
\format{
A 600 by 8 data frame. The true number of classes is 2 based on 7 indicators. The model is
f1 <- cbind(V1, V2, V3, V4, V5, V6, V7) ~ 1. There is an additionnal variable for 3-step testing.
}
\source{
\url{https://github.com/quantmeth}
}
\usage{
ex2.poLCA
}
\description{
A data frame for poLCA and poLCAExtra.
}
\keyword{datasets}
