% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex3.tidyLPA.R
\docType{data}
\name{ex3.tidyLPA}
\alias{ex3.tidyLPA}
\title{An example for poLCAExtra}
\format{
A 300 by 9 data frame. The true number of classes is 3 based on 6 indicators. The model is
f1 <- cbind(V1, V2, V3, V4, V5, V6) ~ 1. There is two additionnal variables, T1, and T2, for 3-step testing and an identication variable.
}
\source{
\url{https://github.com/quantmeth}
}
\usage{
ex3.tidyLPA
}
\description{
A data frame for tidyLPA.
}
\keyword{datasets}
