% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{is_lca_assign}
\alias{is_lca_assign}
\title{lca_assign Class Check}
\usage{
is_lca_assign(x)
}
\arguments{
\item{x}{as object.}
}
\value{
a logical value.
}
\description{
Checks if an object belongs to the \code{lca_assign} class.
}
