% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{is_polcax}
\alias{is_polcax}
\title{polcax Class Check}
\usage{
is_polcax(x)
}
\arguments{
\item{x}{as object.}
}
\value{
a logical value.
}
\description{
Checks if an object belongs to the \code{polcax} class.
}
