% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust_analysis.R
\name{lca2clust.polcax}
\alias{lca2clust.polcax}
\alias{lca2clust.lca_assign}
\alias{lca2clust}
\title{Coerce to a clust_analysis Object}
\usage{
\method{lca2clust}{polcax}(
  x,
  distance_method = c("euclidean", "manhattan", "cosine", "jaccard", "smc"),
  ...
)

\method{lca2clust}{lca_assign}(
  x,
  data,
  distance_method = c("euclidean", "manhattan", "cosine", "jaccard", "smc"),
  ...
)

lca2clust(x, ...)
}
\arguments{
\item{x}{as object.}

\item{distance_method}{distance between the observations. Allowed are:
\code{euclidean} (default), \code{manhattan}, \code{cosine}, \code{jaccard},
\code{smc} (simple matching coefficient) and \code{dice}.}

\item{...}{extra arguments passed to the methods.}

\item{data}{modeling data frame: required only for \code{lca_assign} objects.}
}
\description{
Coerces an object (currently \code{polcax} or \code{lca_assign} class instance) to a
\code{\link[clustTools]{clust_analysis}} object which may be
further investigated with a broad range
of diagnostic and visualization tools provided by
the \href{https://github.com/PiotrTymoszuk/clustTools}{clustTools} package.
}
\details{
\code{lca2clust} is a S3 generic function. Distances between
observations are computed with \code{\link[clustTools]{calculate_dist}}.
}
