% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{lca_assign}
\alias{lca_assign}
\title{lca_assign Class}
\usage{
lca_assign(x, ...)
}
\arguments{
\item{x}{an object. Currently data frames and tibbles are accepted.}

\item{...}{extra arguments, currently none.}
}
\description{
Constructs an instance of the \code{lca_assign} class on the top of a data frame.
}
\details{
The input object needs to be a data frame or a tibble with the following
minimal variable set:
\itemize{
\item \code{observation} with unique observation ID
\item \code{class} a factor with the class identifiers
\item \code{best_p} a numeric value with the probability of the class assignment
}

See also: the \code{plot} method for the \code{lca_assign} class.
}
