% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_methods.R
\name{model.frame.polcax}
\alias{model.frame.polcax}
\title{Modeling Data of the polcax Object}
\usage{
\method{model.frame}{polcax}(formula, ...)
}
\arguments{
\item{formula}{a \code{polcax} class object.}

\item{...}{extra arguments passed to methods, currently none.}
}
\value{
a data frame of matrix with the modeling data.
}
\description{
Retrieves the modeling data of a \code{polcax} object.
}
