% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_methods.R
\name{nobs.polcax}
\alias{nobs.polcax}
\title{Numbers of Observations and Variables}
\usage{
\method{nobs}{polcax}(object, ...)
}
\arguments{
\item{object}{an object of the \code{polcax} class.}

\item{...}{extra arguments passed to methods, currently none.}
}
\value{
a list with the \code{observations} and \code{variables} elements.
}
\description{
Numbers of observations and variables in the \code{polcax} model.
}
