% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{nodal_vote}
\alias{nodal_vote}
\title{Assign Observations to Classes by Nodal Vote.}
\usage{
nodal_vote(x, resolve_ties = FALSE)
}
\arguments{
\item{x}{a matrix with probabilities, observations in rows
and classes in columns.}

\item{resolve_ties}{logical, should ties in vote be resolved by random?
Defaults to \code{FALSE}.}
}
\value{
a tibble with the columns \code{observation}, \code{class} and \code{best_p}.
}
\description{
Assigns observation to classes by a simple nodal probability vote.
}
\details{
The column and rownames of the input matrix need to be specified.
Usually, if there are ties in voting, the first class of the tie is chosen.
Ties will be resolved by random if \code{resolve_ties} is set to \code{TRUE}.
}
