% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.classes.R
\name{plot.classes}
\alias{plot.classes}
\title{Plot results of a model from poLCA2}
\usage{
\method{plot}{classes}(x, ..., nclass = NULL, alpha = 0.05)
}
\arguments{
\item{x}{An object of class "poLCA" or "poLCA2".}

\item{...}{Further Objects of class "poLCA".}

\item{nclass}{If the object is of class "poLCA2" and only a single model have to be plotted.}

\item{alpha}{The type I error rate for confidence intervals.}
}
\value{
A ggplot2 output.
}
\description{
Plot results of a model from poLCA2
}
\examples{
f1 <- as.formula(cbind(V1, V2, V3, V4, V5, V6) ~ 1)
LCA2 <- poLCA(f1, data = ex1.poLCA, nclass = 2)
LCA3 <- poLCA(f1, data = ex1.poLCA, nclass = 3)
plot.classes(LCA2, LCA3)
LCAE <- poLCA(f1, data = ex1.poLCA, nclass = 2:3)
plot.classes(LCAE)
plot.classes(LCAE, nclass = 3)
}
