% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic_methods.R
\name{plot.lca_assign}
\alias{plot.lca_assign}
\title{Graphics for the Latent Class Assignment}
\usage{
\method{plot}{lca_assign}(
  x,
  type = c("class_distribution", "class_posterior"),
  signif_digits = 2,
  txt_size = 2.75,
  txt_color = "black",
  point_size = 2,
  point_alpha = 0.8,
  point_hjitter = 0.005,
  point_wjitter = 0.005,
  flip = FALSE,
  cust_theme = theme_classic(),
  ...
)
}
\arguments{
\item{x}{a \code{lca_assign} class object.}

\item{type}{the requested plot type, see: \code{Details}.}

\item{signif_digits}{significant digits.}

\item{txt_size}{size of the text displayed in the plot labels.}

\item{txt_color}{color of the label text.}

\item{point_size}{point size.}

\item{point_alpha}{point alpha.}

\item{point_hjitter}{height of point jittering.}

\item{point_wjitter}{width of point jittering.}

\item{flip}{logical: flip the plot axes? Defaults to \code{FALSE}.}

\item{cust_theme}{custom ggplot graphic theme.}

\item{...}{extra arguments, currently none.}
}
\value{
a \code{ggplot} object.
}
\description{
Plots class distribution or probability of class assignment.
}
\details{
Two types of plots can be generated as specified by the \code{type} argument:
\itemize{
\item \code{class_distribution} plot class distribution in a stack plot
\item \code{class_posterior} plots probabilities of class assignment for observations
split by the class
}

The last plot type may be useful at assessment of the general
reliability of the model and identification of 'fence-sitter' observations
or classes with with doubtful observation assignment.
}
