% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.poLCA2.R
\name{plot.poLCA2}
\alias{plot.poLCA2}
\title{Plot results of a model from poLCA2}
\usage{
\method{plot}{poLCA2}(x, nclass = NULL, ...)
}
\arguments{
\item{x}{An object of class "poLCA2".}

\item{nclass}{The number of classes determined.}

\item{...}{Further arguments for other methods, ignored for "poLCA2".}
}
\value{
A plot output.
}
\description{
Plot results of a model from poLCA2
}
\examples{
f1 <- as.formula(cbind(V1, V2, V3, V4, V5, V6) ~ 1)
out <- poLCA(f1, nclass = 1:3, data = ex1.poLCA)
plot(out, nclass = 3)
}
