% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic_methods.R
\name{plot.polcax}
\alias{plot.polcax}
\title{Graphics for the polcax Class}
\usage{
\method{plot}{polcax}(
  x,
  type = c("poLCA", "class_distribution", "prior_distribution", "prior_se",
    "class_posterior", "conditional_probs"),
  resolve_ties = FALSE,
  signif_digits = 2,
  txt_size = 2.75,
  txt_color = "black",
  txt_hjust = 0,
  txt_vjust = -1.4,
  point_size = 2,
  point_alpha = 0.8,
  point_hjitter = 0.005,
  point_wjitter = 0.005,
  flip = FALSE,
  cust_theme = ggplot2::theme_classic(),
  ...
)
}
\arguments{
\item{x}{a \code{polcax} object.}

\item{type}{type of the plot or plots, see: \code{Details}.}

\item{resolve_ties}{logical: should ties at class assignment be resolved
at random? Defaults to \code{FALSE}.}

\item{signif_digits}{significant digits.}

\item{txt_size}{size of the text displayed in the plot labels.}

\item{txt_color}{color of the label text.}

\item{txt_hjust}{horizontal justification of the label text.}

\item{txt_vjust}{vertical justification of the label text.}

\item{point_size}{point size.}

\item{point_alpha}{point alpha.}

\item{point_hjitter}{height of point jittering.}

\item{point_wjitter}{width of point jittering.}

\item{flip}{logical: flip the plot axes? Defaults to \code{FALSE}.}

\item{cust_theme}{custom ggplot graphic theme.}

\item{...}{extra arguments, currently none.}
}
\value{
a ggplot object or a list of ggplot objects.
}
\description{
Plots for \code{polcax} class objects.
}
\details{
The method provides a variety of plots to visualize class distribution,
prior, posterior and conditional probabilities
specified by the \code{type} argument:
\itemize{
\item \code{poLCA} generates a traditional \code{poLCA} plot of conditional probabilities
\item \code{class_distribution} plots distribution of the classes expressed as
percentages of all complete observations
\item \code{prior_distribution} plots distribution of the prior probabilities
\item \code{prior_se} plots prior p values with their standard errors
\item \code{class_posterior} plots probabilities of class assignment for observations
split by the class
\item \code{conditional_probs} returns a list of plots with conditional
probabilities of the class assignment for each modeling variable
}
}
