% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r3step.R
\name{plot.r3step}
\alias{plot.r3step}
\title{Plot of r3step, bootstraped 3-step test of latent class analysis}
\usage{
\method{plot}{r3step}(x, ci = NULL, ...)
}
\arguments{
\item{x}{An object of class "3step".}

\item{ci}{Confidence intervals (if they have to be changed).}

\item{...}{Further arguments for other methods, ignored for "r3step".}
}
\value{
A ggplot output.
}
\description{
Plot of r3step, bootstraped 3-step test of latent class analysis
}
\examples{
jd <- ex1.poLCA
f1 <- cbind(V1, V2, V3, V4, V5, V6) ~ 1
LCAE <- poLCA(f1, jd, nclass = 1:4, nrep = 4, maxiter = 1000)
results <- r3step("continuous", LCAE, nclass = 3)
plot(results)
}
