% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poLCA.blrt.R
\name{poLCA.blrt}
\alias{poLCA.blrt}
\title{Bootstrapped likelihood ratio tests}
\usage{
poLCA.blrt(object, nreps = 50, ...)
}
\arguments{
\item{object}{an object of class "poLCA" or "poLCA2".}

\item{nreps}{the number of replications for the bootstrap. Default is 50.}

\item{...}{any other argument for \code{poLCA}.}
}
\value{
A data frame containing mainly :
\itemize{
\item Average bootstraped Vuong-Lo-Mendell-Rubin adjusted likelihood ratio test;
\item Average bootstraped Lo-Mendell-Rubin adjusted likelihood ratio test;
\item Vuong-Lo-Mendell-Rubin adjusted likelihood ratio test p-value;
\item Lo-Mendell-Rubin adjusted likelihood ratio test p-value.}
}
\description{
Bootstrapped likelihood ratio tests
}
\details{
This function can take some time especially with "poLCA2" class objects as they contain many "poLCA" objects.
}
\examples{
f1 <- cbind(V1, V2, V3, V4, V5, V6) ~ 1
out <- poLCA(f1, nclass = 3:4, data = ex1.poLCA)
poLCA.blrt(out)
}
\references{
Linzer, D. A. & Lewis, J. F. (2011). poLCA: An R Package for Polytomous Variable Latent Class Analysis. \emph{Journal of Statistical Software}, \emph{42}(10), 1-29. \url{https://www.jstatsoft.org/v42/i10/}
}
\author{
P.-O. Caron
}
