% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poLCA.compare.R
\name{poLCA.compare}
\alias{poLCA.compare}
\title{Compare the statistics of an poLCA2 object}
\usage{
poLCA.compare(x, stat = "AIC", alpha = 0.05)
}
\arguments{
\item{x}{an poLCA2 object.}

\item{stat}{the desired statistics to compare. Default is "AIC". It can be also "BIC", "saBIC", "CAIC" or "AIC3".}

\item{alpha}{the significance threshold.}
}
\value{
A data frame with 
\itemize{
  \item \code{The classes comparison} - which nclass are compared.
  \item \code{chisq} - The chi square.
  \item \code{df} - The degree of freedom.
  \item \code{p} - The p-value.
  \item \code{remark} - A note indicating if it a significant decrease, a significant increase or no statistical difference based on \code{alpha}.
}
}
\description{
Compare the statistics of an poLCA2 object
}
\examples{
f1 <- cbind(V1, V2, V3, V4, V5, V6) ~ 1
LCAE <- poLCA(f1, nclass = 1:3, data = ex1.poLCA)
poLCA.compare(LCAE)
}
