% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poLCA.cov.R
\name{poLCA.cov}
\alias{poLCA.cov}
\title{Inspect the model covariance matrix and some fit statistics}
\usage{
poLCA.cov(x, nclass = NULL, ...)
}
\arguments{
\item{x}{object of class inheriting from "poLCA" or "poLCA2".}

\item{nclass}{if \code{object} contains many "poLCA" object (i.e., if it is a "poLCA2" object), choose the model with the given number of classes.}

\item{...}{further objects of class "poLCA".}
}
\value{
A list of covariances and some relevant statistics.
}
\description{
Inspect the model covariance matrix and some fit statistics
}
\examples{
f1 <- cbind(V1, V2, V3, V4, V5, V6) ~ 1
out <- poLCA(f1, nclass = 3, data = ex1.poLCA)
poLCA.cov(out)
}
