% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poLCA.lmr.R
\name{poLCA.lmr}
\alias{poLCA.lmr}
\title{Lo-Mendell-Rubin adjusted likelihood ratio test}
\usage{
poLCA.lmr(model1, model2 = NULL)
}
\arguments{
\item{model1}{First poLCA model.}

\item{model2}{Second poLCA model to compare to the first one.}
}
\value{
A list containing the Vuong-Lo-Mendell-Rubin adjusted likelihood ratio test, the Lo-Mendell-Rubin adjusted likelihood ratio, the degree of freedom and the lmr's p-value.
}
\description{
Lo-Mendell-Rubin adjusted likelihood ratio test
}
\examples{
f1 <- as.formula(cbind(V1, V2, V3, V4, V5, V6) ~ 1)
LCA2 <- poLCA(f1, nclass = 2, data = ex1.poLCA)
LCA3 <- poLCA(f1, nclass = 3, data = ex1.poLCA)

poLCA.lmr(LCA3, LCA2)
# or
poLCA.lmr(LCA3)
}
