% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poLCA.plot.R
\name{poLCA.plot}
\alias{poLCA.plot}
\title{Plot results of a model from poLCA2}
\usage{
poLCA.plot(x, ..., nclass = NULL, alpha = 0.05)
}
\arguments{
\item{x}{An object of class "poLCA" or "poLCA2".}

\item{...}{Further objects of class "poLCA".}

\item{nclass}{If the object is of class "poLCA2" and only a single model have to be plotted.}

\item{alpha}{The type I error rate for confidence intervals.}
}
\value{
A ggplot2 output.
}
\description{
Plot results of a model from poLCA2
}
\examples{
f1 <- as.formula(cbind(V1, V2, V3, V4, V5, V6) ~ 1)
LCA2 <- poLCA(f1, data = ex1.poLCA, nclass = 2)
LCA3 <- poLCA(f1, data = ex1.poLCA, nclass = 3)
poLCA.plot(LCA2, LCA3)
LCAE <- poLCA(f1, data = ex1.poLCA, nclass = 2:3)
poLCA.plot(LCAE)
poLCA.plot(LCAE, nclass = 3)
}
