% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poLCA.R
\name{poLCA}
\alias{poLCA}
\title{Testing several latent class analyses of polytomous outcome variables}
\usage{
poLCA(formula, data, nclass = NULL, verbose = FALSE, ...)
}
\arguments{
\item{formula}{A formula expression of the form response ~ predictors. The details of model specification are given below.}

\item{data}{A data frame containing variables in formula. Manifest variables must contain only integer values, and must be coded with consecutive values from 1 to the maximum number of outcomes for each variable. All missing values should be entered as NA.}

\item{nclass}{A vector containing the number of latent classes to assume in the models. Setting nclass=1 results in poLCA estimating the loglinear independence model. Default is 1 to the maximum number of classes possible.}

\item{verbose}{Logical, indicating whether poLCA should output to the screen the results of the model. If FALSE, no output is produced. The default is FALSE.}

\item{...}{Any other argument for \code{poLCA}.}
}
\value{
\code{poLCA} returns an object of class "poLCA" if \code{length(nclass) == 1} or returns an object of class "poLCA2" if \code{length(nclass) > 1}; a list containing the following elements :

\itemize{
  \item \code{output} - A summary table of the fit.
  \item \code{data} - The data used.
  \item \code{LCA} - A list of objects of class "poLCA" (\code{poLCA} model)
}
}
\description{
#' \code{poLCA} is an adaptation of the eponym function in the package \code{poLCA} to account for many models simultaneously
}
\examples{
f1 <- cbind(V1, V2, V3, V4, V5, V6) ~ 1
poLCA(f1, nclass = 1:3, data = ex1.poLCA)
}
\references{
Linzer, D. A. & Lewis, J. F. (2011). poLCA: An R Package for Polytomous Variable Latent Class Analysis. \emph{Journal of Statistical Software}, \emph{42}(10), 1-29. \url{https://www.jstatsoft.org/v42/i10/}
}
\author{
P.-O. Caron
}
