% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poLCA.relentropy.R
\name{poLCA.relentropy}
\alias{poLCA.relentropy}
\title{Compute relative entropy of an "poLCA" object}
\usage{
poLCA.relentropy(model)
}
\arguments{
\item{model}{An object of class "poLCA".}
}
\value{
A vector containing the relative entropy.
}
\description{
Compute relative entropy of an "poLCA" object
}
\examples{
f1 <- as.formula(cbind(V1, V2, V3, V4, V5, V6) ~ 1)
out <- poLCA(f1, nclass = 3, data = ex1.poLCA)
poLCA.relentropy(out)
}
