% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poLCA.residualcov.R
\name{poLCA.residual.cov}
\alias{poLCA.residual.cov}
\title{Inspect residual covariances of each class.}
\usage{
poLCA.residual.cov(x, nclass = NULL)
}
\arguments{
\item{x}{Object of class inheriting from "poLCA" or "poLCA2".}

\item{nclass}{If \code{object} contains many "poLCA" object (i.e., if it a "poLCA2" object), choose the model with the given number of classes.}
}
\value{
A list of residual covariance matices.
}
\description{
Inspect residual covariances of each class.
}
\examples{
f1 <- cbind(V1, V2, V3, V4, V5, V6) ~ 1
out <- poLCA(f1, nclass = 3, data = ex1.poLCA)
poLCA.residual.cov(out)
}
