% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tech10.R
\name{poLCA.residual.pattern}
\alias{poLCA.residual.pattern}
\alias{poLCA.tech10}
\title{Inspect patterns' residuals of latent class analysis.}
\usage{
poLCA.residual.pattern(x, nclass = NULL, alpha = 0.05)

poLCA.tech10(x, nclass = NULL, alpha = 0.05)
}
\arguments{
\item{x}{an object of class "poLCA" or "poLCA2"}

\item{nclass}{if \code{object} contains many "poLCA" object (i.e., if it a "poLCA2" object), choose the model with the given number of classes.}

\item{alpha}{a type I error threshold to mark the check column.}
}
\value{
A data frame containing the patterns' residuals.
}
\description{
Inspect patterns' residuals of latent class analysis.
}
\examples{
f1 <- as.formula(cbind(V1, V2, V3, V4, V5, V6) ~ 1)
out <- poLCA(f1, nclass = 3, data = ex1.poLCA)
poLCA.residual.pattern(out)
}
