% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poLCA.rmsea.R
\name{poLCA.rmsea}
\alias{poLCA.rmsea}
\title{RMSEA for poLCA objects}
\usage{
poLCA.rmsea(object)
}
\arguments{
\item{object}{an object of class "poLCA or "poLCA2"}
}
\value{
A vector containting RMSEA value of each model
}
\description{
RMSEA for poLCA objects
}
\examples{
f2 <- cbind(V1, V2, V3, V4, V5, V6, V7) ~ 1
LCAE <- poLCA(f2, nclass = 1:4, data = ex2.poLCA)
poLCA.rmsea(LCAE)
}
