% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{polcax}
\alias{polcax}
\alias{as_polcax}
\alias{as_polcax.poLCA}
\title{polcax Class}
\usage{
polcax(x, class_names = NULL, ...)

as_polcax(x, ...)

\method{as_polcax}{poLCA}(x, class_names = NULL, ...)
}
\arguments{
\item{x}{and object used to build the \code{polcax} class instance.
Currently only \code{poLCA} models are accepted.}

\item{class_names}{an optional argument specifying custom class names.
If \code{NULL}, class names are created with the 'class_1' ... 'class_n' scheme.}

\item{...}{arguments for the methods, currently none.}
}
\value{
an object of the \code{polcax} class, which inherits from
the genuine \code{poLCA} models.
}
\description{
Constructs an instance of the \code{polcax} class, which extends the
diagnostic and graphical interface for latent class models fitted with
\code{\link[poLCA]{poLCA}}.
}
\details{
The output object contains the usual components of the
\code{\link[poLCA]{poLCA}} class and the \code{class_names} vector, which stores
the class names.
The input LCA models hast to have named observations, i.e. the modeling data
frame or matrix in the \code{\link[poLCA]{poLCA}} model
has to have row names.
}
