% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.poLCA2.R
\name{predict.poLCA2}
\alias{predict.poLCA2}
\title{Predict method for poLCA and poLCA2 objects}
\usage{
\method{predict}{poLCA2}(object, nclass = NULL, ...)
}
\arguments{
\item{object}{Object of class inheriting from "poLCA" or "poLCA2".}

\item{nclass}{If \code{object} contains many "poLCA" object (i.e., if it a "poLCA2" object), choose the model with the given number of classes.}

\item{...}{Additional arguments affecting the predictions produced.}
}
\value{
A data frame with the probabilities of each subject to be in each classes and the predicted classes.
}
\description{
Get probabilities of each subject to be in a given class and the most likely class.
}
\examples{
f1 <- cbind(V1, V2, V3, V4, V5, V6) ~ 1
out <- poLCA(f1, nclass = 1:3, data = ex1.poLCA)
predict(out, nclass = 3)
}
