% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_methods.R
\name{predict.polcax}
\alias{predict.polcax}
\title{Predict Class Assignment}
\usage{
\method{predict}{polcax}(object, newdata = NULL, x = NULL, resolve_ties = FALSE, ...)
}
\arguments{
\item{object}{a \code{polcax} class object.}

\item{newdata}{a vector or matrix containing series of responses on
the manifest variables in the LCA model.}

\item{x}{an optional vector or matrix of covariate values.}

\item{resolve_ties}{should ties in class assignment prediction
be randomly resolved? Defaults to \code{FALSE}.}

\item{...}{extra arguments, currently none.}
}
\value{
a tibble of class \code{lca_assign} with
the \code{\link{plot.lca_assign}} method specified.
}
\description{
Predicts class assignment based on the LCA model information stored in the
\code{polcax} object.
}
\details{
See: \code{\link[poLCA]{poLCA.posterior}} for details of prediction.
The class assignment is done by voting as specified
for \code{\link{nodal_vote}}.
}
