% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.poLCA2.R
\name{print.poLCA2}
\alias{print.poLCA2}
\title{Print results of poLCA2 objects}
\usage{
\method{print}{poLCA2}(x, digit = 3, ...)
}
\arguments{
\item{x}{Print the table comparing latent class models.}

\item{digit}{integer indicating the number of decimal places.}

\item{...}{Further arguments for other methods, ignored for "poLCA2".}
}
\value{
No return value, called for side effects.
}
\description{
Print results of poLCA2 objects
}
\examples{
f1 <- as.formula(cbind(V1, V2, V3, V4, V5, V6) ~ 1)
poLCA(f1, nclass = 1:3, data = ex1.poLCA)
}
