% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.poLCA2.R
\name{summary.poLCA2}
\alias{summary.poLCA2}
\title{Summary method for "poLCA2" objects}
\usage{
\method{summary}{poLCA2}(object, ...)
}
\arguments{
\item{object}{An object of class "poLCA2".}

\item{...}{Further arguments for other methods, ignored for "poLCA2".}
}
\value{
Return a table with relevant.
}
\description{
Summary method for "poLCA2" objects
}
\note{
Should not be called directly
}
