% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuning.R
\name{tune_lca}
\alias{tune_lca}
\title{Tune LCA Models.}
\usage{
tune_lca(
  formula,
  data,
  class_range = 1:3,
  metric = c("none", "aic", "bic"),
  seed = NULL,
  color = "steelblue",
  size = 2,
  cust_theme = theme_classic(),
  ...
)
}
\arguments{
\item{formula}{a model formula, see: \code{\link[poLCA]{poLCA}}
for specification.}

\item{data}{a modeling data frame.}

\item{class_range}{a vector of integers with the requested class number
to be tested.}

\item{metric}{tuning metric used for search of the optimal model:
\code{none} (default), \code{aic} or \code{bic}. If not \code{none}, the optimal model will be
indicated in the statistic output and in the plots.}

\item{seed}{a numeric vector for the seed of the random number generator.
If \code{NULL} (default), no seed is set.}

\item{color}{color of lines and points in the plots.}

\item{size}{size of points in the plots.}

\item{cust_theme}{custom ggplot theme for the plots.}

\item{...}{extra arguments passed to \code{\link[poLCA]{poLCA}}.}
}
\value{
a list with the following elements:
\itemize{
\item \code{models} a list of \code{\link[poLCA]{poLCA}} models with the requested
class numbers. The models are wrapped as \code{polcax} class instances
\item \code{stats} a data frame with fit statistics and indicators of model
convergence and validity, see: \code{\link{summary.polcax}} for details
\item \code{plots} a list of ggplot drawings visualizing log-likelihood,
AIC (Akaike Information Criterion), BIC (Bayesian Information Criterion),
deviance ratios and chi-squared statistics for the requested class
number range
}
}
\description{
Constructs a series of \code{poLCA} models and compares their convergence,
validity and fit stats.
}
\details{
The function fits a series of \code{\link[poLCA]{poLCA}} models
for a range of class numbers.
Parallel backend can be provided via \code{\link[future]{plan}}.
}
