% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.poLCA.R
\name{anova.poLCA}
\alias{anova.poLCA}
\title{Compare multiple poLCA output}
\usage{
\method{anova}{poLCA}(object, ...)
}
\arguments{
\item{object}{An object of class "poLCA".}

\item{...}{additional objects of the same type.}
}
\value{
A table comparing the poLCA models.
}
\description{
Compare multiple poLCA output
}
\examples{
f1 <- as.formula(cbind(V1, V2, V3, V4, V5, V6) ~ 1)
LCA1 <- poLCA::poLCA(f1, data = jd, nclass = 1) 
LCA2 <- poLCA::poLCA(f1, data = jd, nclass = 2)
LCA3 <- poLCA::poLCA(f1, data = jd, nclass = 3)
anova(LCA3, LCA2, LCA1)
}
