% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.poLCA2.R
\name{summary.poLCA2}
\alias{summary.poLCA2}
\title{Summary method for "poLCA2" objects}
\usage{
\method{summary}{poLCA2}(object, ...)
}
\arguments{
\item{object}{An object of class "poLCA2".}

\item{...}{Further arguments for other methods, ignored for "poLCA2".}
}
\value{
Return a table with relevant
}
\description{
Summary method for "poLCA2" objects
}
\examples{
f1 <- as.formula(cbind(V1, V2, V3, V4, V5, V6) ~ 1)
out <- poLCA(f1, nclass = 1:3, data = ex1.poLCA)
summary(out)
}
