% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.poLCA2.R
\name{summary.poLCA2}
\alias{summary.poLCA2}
\title{Summary results of object "poLCA2"}
\usage{
\method{summary}{poLCA2}(object, alpha = 0.05, ...)
}
\arguments{
\item{object}{an object of class "poLCA2".}

\item{alpha}{type I error rate. Default is .05.}

\item{...}{additional arguments affecting the summary produced.}
}
\value{
A data.frame containing the number of classes identified by different fit statistics.
}
\description{
Summary results of object "poLCA2"
}
\examples{
f2 <- cbind(V1, V2, V3, V4, V5, V6, V7) ~ 1
LCAE <- poLCA(f2, nclass = 1:4, data = ex2.poLCA)
summary(LCAE)
}
